backend/requirements.txt
fastapi
uvicorn
pymongo
transformers
python-dotenv
torch
python-multipart
